[Remove]invShareCommandを削除
